requirements.txt
pywhatkit==5.3
python-telegram-bot==20.7
